break language selector out into its own component as per @danieldietzler suggestion
